trim data if time values do not match
This seems to be a new bug in the data that came up around the end of
Sept, 2020.  My fix will have no effect if sub$time and e.g. sub$cases
are of equal length, but if they differ, it trims trailing data from the
later.  I'll check again in a week or so, to see if the trimming is
still taking place and, if so, I'll investigate further.