perhaps improve plot accuracy BUT
I am just guessing on a lot of the data structure.  For example, I
notice something called "update" and thought that would let us find the
days when data get posted.  Well, maybe, and maybe not; there are also a
lot of NAs in that stream.

I removed the Canada plot because the numbers don't add up -- it was
stating that the overall rate was lower than the rate in any province.
Is something wrong?  Should I have special code for that case?  And will
any time I invest in guessing actually pay off, since the dataset can
change at any time?

All I can say is that the NS data are *roughly* in line with what they
have on the website.  However, the "cases in last 7 days" they list is
not the same as any number in the dataset.  So, really, it's all a
confusing mess. Covid fog in data.